+++ IOS10-HW14-Maria Glushtsova/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
