+++ IOS10-HW14-Maria Glushtsova/Resources/Assets.xcassets/Contents.json
