+++ IOS10-HW14-Maria Glushtsova/Sources/App/AppDelegate.swift
@@ -1,0 +1,17 @@
+//
+//  AppDelegate.swift
+//  IOS10-HW14-Maria Glushtsova
+//
+//  Created by Admin on 13.07.23.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        // Override point for customization after application launch.
+        return true
+    }
+}
+
