+++ IOS10-HW14-Maria Glushtsova/Sources/App/SceneDelegate.swift
@@ -1,0 +1,22 @@
+//
+//  SceneDelegate.swift
+//  IOS10-HW14-Maria Glushtsova
+//
+//  Created by Admin on 13.07.23.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        window = UIWindow(windowScene: windowScene)
+        let viewController = ViewController()
+        let navigationController = UINavigationController(rootViewController: viewController)
+        window?.rootViewController = navigationController
+        window?.makeKeyAndVisible()
+    }
+}
